Map Python exception to equivalent Nim exception, if possible. Closes #18 (#56)
* add mapping of Python exceptions to Nim exceptions

Add `PythonErrorKind` enum, which maps a Python exception via the
string value of the fields to the corresponding Nim exception. The
field names are exactly the names of the Nim exception plus a "pe"
prefix. This is used to build the case statement with a macro.

* fix strscans string for Python 2

* add test to check exception mapping

* replace try/except by `parseEnum` with default

* add proper exception mapping, aware of derived Exception classes

Adds exception mapping of Python exceptions to Nim exceptions that is
aware of derived Python exceptions.

Using `PyErr_GivenExceptionMatches` we check for all Python
exceptions, we can map to Nim exceptions. We call for each Python
exception we load, from general to specific if
applicable. E.g. we'd check `LookupError` before `IndexError`, because
`IndexError` is derived from `LookupError` and both match.

This allows us to match even derived exceptions (from the exceptions
we load anyways).

* clean up / fix comment

* remove `LookupError` Python exception

We have no equivalent in Nim at the moment. It was added for debugging
to check whether derived exceptions are properly catched.